Extracting Type of Advice filter from search form
